+++ internal/control_plane/command/start/start.go
@@ -3,7 +3,7 @@ package start
 import (
 	"fmt"
 	"log"
-	"octavius/internal/controlPlane/server"
+	"octavius/internal/control_plane/server"
 
 	"github.com/spf13/cobra"
 )
